requirements.txt
# ============================================================
# requirements.txt
# 목적: 런타임 의존성 (FastAPI 앱 실행에 필요한 패키지만 포함)
# 적용 환경: Docker 컨테이너, 운영 환경
# 설치 방법: pip install -r requirements.txt
#
# [주의]
# - 개발/테스트 도구(flake8, black, pytest)는 포함하지 않음
# - Docker 이미지 경량화를 위해 런타임 패키지만 명시
# - 버전 고정(==)으로 CI 및 운영 환경 재현성 보장
# ============================================================

# HTTP 요청 클라이언트 — APIC REST API 호출에 사용
requests==2.32.3

# YAML 파서 — config.yaml (APIC 접속 정보) 로드에 사용
pyyaml==6.0.2

# FastAPI 웹 프레임워크 — API 서버 구현
fastapi==0.115.12

# ASGI 서버 — FastAPI 앱 실행 (uvicorn main:app)
uvicorn==0.34.0

# 비동기 HTTP 클라이언트 — FastAPI TestClient 의존성 (pytest 테스트에 필요)
httpx==0.28.1

# multipart 폼 데이터 파싱 — 파일 업로드 처리에 사용
python-multipart==0.0.20

# JWT 생성/검증 — 로그인 토큰 발급 및 인증에 사용 (v1.9.2)
pyjwt==2.10.1

# 비밀번호 해싱 — bcrypt 기반 사용자 비밀번호 저장에 사용 (v1.9.2)
passlib[bcrypt]==1.7.4